+++ dds_cli/auth.py
@@ -38,8 +38,8 @@ class Auth(base.DDSBaseClass):
         token_file = user.TokenFile()
         if token_file.file_exists():
             token_file.check_token_file_permissions()
-            _, token_metadata = token_file.read_token()
-            token_file.token_report(token_metadata=token_metadata)
+            token = token_file.read_token()
+            token_file.token_report(token=token)
         else:
             LOG.error(f"[red]No saved authentication token found![/red]")