+++ CHANGELOG.md
@@ -82,3 +82,4 @@ Please add a _short_ line describing the PR you make, if the PR implements a spe
 * Save failed files to log and print out help message after ([237](https://github.com/ScilifelabDataCentre/dds_cli/pull/237))
 * Change `--is_sensitive` to `--non-sensitive` ([#246](https://github.com/ScilifelabDataCentre/dds_cli/pull/246))
 * Display logged in user in header ([244](https://github.com/ScilifelabDataCentre/dds_cli/pull/244))
+* Updated token expiration information ([245](https://github.com/scilifelabdatacentre/dds_cli/issues/245))